+++ ezomero/__init__.py
@@ -1,6 +1,5 @@
 from ._ezomero import (put_map_annotation,
                        connect,
-                       create_json_session,
                        store_connection_params,
                        set_group)
 from ._misc import (filter_by_filename,
@@ -35,7 +34,6 @@ from ._gets import (get_image,
                     get_user_id,
                     get_original_filepaths,
                     get_pyramid_levels,
-                    get_rendered_jpeg,
                     get_shape)
 
 __all__ = ['post_dataset',
@@ -60,7 +58,6 @@ __all__ = ['post_dataset',
            'get_user_id',
            'get_original_filepaths',
            'get_pyramid_levels',
-           'get_rendered_jpeg',
            'get_shape',
            'put_map_annotation',
            'filter_by_filename',
@@ -74,6 +71,5 @@ __all__ = ['post_dataset',
            'print_projects',
            'print_datasets',
            'connect',
-           'create_json_session',
            'store_connection_params',
            'set_group']
