+++ ezomero/_gets.py
@@ -1,8 +1,6 @@
 import logging
 import os
 import numpy as np
-from PIL import Image
-from io import BytesIO
 from ._ezomero import do_across_groups
 from omero.gateway import FileAnnotationWrapper
 from omero import ApiUsageException
@@ -962,58 +960,6 @@ def get_pyramid_levels(conn, image_id, across_groups=True):
     levels = [(r.sizeX, r.sizeY) for r in pix.getResolutionDescriptions()]
     pix.close()
     return levels
-
-def get_rendered_jpeg(session, base_url, img_id, scale):
-    """Get a numpy array from a rendered JPEG at given scale factor
-    of an image.
-
-    Parameters
-    ----------
-    session : ``requests`` Session object or None
-        ``requests`` session that has been initialized with login parameters.
-    base_url : str or None
-        Base URL for further requests, retrieved via JSON API request
-    img_id : int
-        ID of ``Image``.
-    scale : float
-        Scaling factor for the returned JPEG. ``1`` returns original size,
-        ``2`` scales each dimension down by half, and so on.
-
-    Returns
-    -------
-    pixels : ndarray
-        Array containing pixel values from the rendered JPEG of the OMERO image. 
-
-    Examples
-    --------
-    # Get a JPEG of an entire image as a numpy array:
-
-    >>> jpeg_array = get_rendered_jpeg(session, base_url, 314, 1)
-
-    """
-    import numpy as np
-    # just some magical code to get the correct address from the json api session and image id
-    r = session.get(base_url)
-    host = base_url.split("/api")[0]
-    # which lists a bunch of urls as starting points
-    urls = r.json()
-    images_url = urls['url:images']
-    single_image_url = images_url+str(img_id)+"/"
-    thisjson = session.get(single_image_url).json()
-
-    # calculate width to be requested based on metadata and the specified scale factor
-    width = int(thisjson['data']['Pixels']['SizeX'])
-    scaled = round(width/scale)
-    img_address = host+"/webgateway/render_birds_eye_view/"+str(img_id)+"/"+str(scaled)+"/"
-    jpeg = session.get(img_address, stream=True)
-
-    if jpeg.status_code != 200:
-        raise Exception("Received response {} with content: {}".format(jpeg.status_code, jpeg.content))
-
-    # using PIL and BytesIO to open the request content as an image
-    i = Image.open(BytesIO(jpeg.content))
-    jpeg.close()
-    return np.array(i)
 
 
 @do_across_groups
