+++ tests/test_json_api.py
@@ -1,0 +1,4 @@
+import pytest
+from ezomero import json_api
+
+
